Create and insert index keys into db
This commit creates and inserts index keys into the database.

We also alter encode_tagged_value() to take a new TaggableValue enum,
which we use for both the path components and the JSON values. It
maps all JSON primitive values along with the string field names and
usize array indexes into a single enum, and we encode them all as
tagged values alongside the docid. This is unlike the original Go code
where we only tagged the value as we knew the other types were strings
in the [path, value, docid] key.

In the end, I think that using tagged values for everything should
allow us to make the encoding/decoding of the key more general
(essentially a null-separated array of tagged values) and interpret
the resulting vectors of decoded TaggedValues that we will need when
executing queries and extracting the docids from the keys as we go
through the ranges.

We will see if it ends up simpler or not!