Wrap insert and delete document in transactions
get_document doesn't need a transaction as it
is a single atomic get on the database.

Also searching the index doesn't yet happen in
a transaction; it likely should.

Alternatively, we could revert to batches and
a single read-write lock, but I think that Sled's
transactions are likely the better choice.