Refactor database range scan to doc IDs code into scan() function
While adding a function, I think this decreases complexity because it
shows what each lookup_* function's unique contribution is --- creating
the right start and end keys for the range scan.

This will simplify moving the search_index function to use a
transaction, which it needs to if it is to avoid inconsistent
results in the face of concurrent document updates.

This should help too if and when we start to optimise the query ---
for example, if we have a GTE and LT for the same field that can
be combined into a single range scan. It helps because it shows us
that we need only figure out how to create the approporiate start and
end keys during the optimisation step.