Revert insert/delete to batched; add updating docs
Sled transactions don't allow range queries, so searches couldn't
be done in a transaction over a snapshot of the database because they
may execute multiple range queries.

So if we want to be thread safe, we need to implement a simple
read/write lock over the get/insert/delete/search functions. This
will mean that writes cannot happen concurrently with any reads,
but so be it.