Alter lookup* to return start/end key rather than ids
Use the start/end key returned to call scan() in the loop within
search_index. This lays the groundwork for us to gather together
the start/end keys grouped by field, then work out the smallest range
scan that covers the key ranges for a given field to collapse down
multiple predicates for the same key into a single scan.