Fail in query planning for inconsistent predicates
In the query planner, we group predicates by their field path and
try to construct a single range scan that will satisfy every predicate.
If we cannot do that, then the predicates must be of the form
"a < 12 AND a > 12" so no document can satisfy them.

This commit adds code to recognise when there is no range that satisfies
all predicates, and adds tests for both successful and unsuccessful
query planning.